% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_model.R
\name{compare_model}
\alias{compare_model}
\title{Test for significant difference in power between a model and a simpler
submodel.}
\usage{
compare_model(fit = NULL, nsim = 1000, seed = NULL)
}
\arguments{
\item{fit}{lme4 object}

\item{nsim}{number of simulations to perform}

\item{seed}{random number generator seed, for reproducible results}
}
\value{
Power analysis results.
}
\description{
Test for significant difference in power between a model and a simpler
submodel.
}
