% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\name{get_model}
\alias{get_model}
\title{Obtain a model from supplied parameters.}
\usage{
get_model(
  fixed_effects = NULL,
  random_effects = NULL,
  effect_size = NULL,
  effect_sd = NULL,
  response = NULL,
  family = NULL,
  size = NULL
)
}
\arguments{
\item{fixed_effects}{List of names of fixed effects.}

\item{random_effects}{List of names of random effects.}

\item{effect_size}{Effective sample size.}

\item{effect_sd}{Effective sample standard deviation.}

\item{response}{Response variable}

\item{family}{Family of model}

\item{size}{Number of observations}
}
\value{
model with supplied parameters
}
\description{
Obtain a model from supplied parameters.
}
