% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_power.R
\name{get_power}
\alias{get_power}
\title{Obtain the power analysis on a model.}
\usage{
get_power(fit = NULL, nsim = 1000, seed = NULL)
}
\arguments{
\item{fit}{A fitted model object.}

\item{nsim}{number of simulations to perform}

\item{seed}{random number generator seed, for reproducible results}
}
\value{
Power analysis results.
}
\description{
Obtain the power analysis on a model.
}
