% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_power.R
\name{get_power}
\alias{get_power}
\title{Obtain the power analysis on a model.}
\usage{
get_power(
  fixed_effects = NULL,
  random_effects = NULL,
  effect_size = NULL,
  effect_variance = NULL
)
}
\arguments{
\item{fixed_effects}{List of names of fixed effects.}

\item{random_effects}{List of names of random effects.}

\item{effect_size}{Effective sample size.}

\item{effect_variance}{Effective sample variance.}
}
\value{
Power analysis results.
}
\description{
Obtain the power analysis on a model.
}
